float64 weights

+++ detectors/model_v5.py
@@ -5,8 +5,14 @@ from model_base import ModelBase
 
 def create_conv(in_tensor, kernel, out_features, activation='relu'):
   in_features = in_tensor.get_shape()[3].value
-  w_c = tf.Variable(tf.random_normal([kernel, kernel, in_features, out_features]))
-  b_c = tf.Variable(tf.random_normal([out_features]))
+  w_c = tf.Variable(tf.random_normal(
+    [kernel, kernel, in_features, out_features],
+    dtype=tf.float64,
+  ))
+  b_c = tf.Variable(tf.random_normal(
+    [out_features],
+    dtype=tf.float64,
+  ))
   conv = tf.nn.conv2d(in_tensor, w_c, strides=[1, 1, 1, 1], padding='SAME')
   conv_w_bias = tf.add(conv, b_c)
   return activation_functions[activation](conv_w_bias)
